Disable the cache before every reload
